perf: lazy load Teleport component
