+++ apps/web/src/routes/index.tsx
@@ -8,6 +8,7 @@ import AccountValueInfo from '@components/recipes/AccountValueInfo'
 import AccountConnectionGuard, { useShouldShowAccountConnectionGuard } from '@components/widgets/AccountConnectionGuard'
 import AccountsManagementMenu from '@components/widgets/AccountsManagementMenu'
 import { RouteErrorElement } from '@components/widgets/ErrorBoundary'
+import TeleportDialog from '@components/widgets/TeleportDialog'
 import StakeDialog from '@components/widgets/staking/StakeDialog'
 import { selectedAccountsState } from '@domains/accounts/recoils'
 import * as MoonbeamContributors from '@libs/moonbeam-contributors'
@@ -28,14 +29,12 @@ import { usePostHog } from 'posthog-js/react'
 import { useCallback, useEffect, useState } from 'react'
 import { Link, Navigate, Outlet, createBrowserRouter, useLocation, useSearchParams } from 'react-router-dom'
 import { useRecoilValue } from 'recoil'
-
-import TeleportDialog from '@components/widgets/TeleportDialog'
 import AssetItem from './AssetItem'
 import Assets from './Assets'
+import Collectibles from './Collectibles'
 import CrowdloanDetail from './Crowdloan.Detail'
 import CrowdloanIndex from './Crowdloan.Index'
 import Explore from './Explore'
-import Collectibles from './Collectibles'
 import Overview from './Overview'
 import Portfolio from './Portfolio'
 import TransactionHistory from './TransactionHistory'
